requirements.txt
opencv-python>=4.8.0
numpy>=1.25.0
Pillow>=10.0.0